Add more utest case for DTLS client.
